extract telegram response handlers
